Introduce ApplicativeOrderEvaluationInterpreter
