+++ src/index.ts
@@ -10,7 +10,7 @@ import {
   UndefinedVariable
 } from './errors/errors'
 import { RuntimeSourceError } from './errors/runtimeSourceError'
-import { evaluate } from './interpreter/interpreter'
+import { ApplicativeOrderEvaluationInterpreter } from './interpreter/interpreter'
 import { parse, parseAt } from './parser/parser'
 import { AsyncScheduler, PreemptiveScheduler } from './schedulers'
 import { getAllOccurrencesInScope, lookupDefinition, scopeVariables } from './scoped-vars'
@@ -249,7 +249,8 @@ export async function runInContext(
       )
     }
   } else {
-    const it = evaluate(program, context)
+    const interpreter = new ApplicativeOrderEvaluationInterpreter()
+    const it = interpreter.evaluate(program, context)
     let scheduler: Scheduler
     if (theOptions.scheduler === 'async') {
       scheduler = new AsyncScheduler()
